# api-gateway/api/docs/swagger.yaml
basePath: /v1
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.ForgotPasswordRequest:
    properties:
      email:
        type: string
    type: object
...
  models.GetListStaffResponse:
    properties:
      count:
        type: integer
      staff:
        items:
          $ref: '#/definitions/models.Staff'
        type: array
    type: object
  models.LogInRequest:
    properties:
      email:
        type: string
      passward:
        type: string
    type: object
  models.ResponseOK:
    properties:
      message:
        type: string
    type: object
...
  models.Staff:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      image_url:
        type: string
      last_name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      salon_id:
        type: string
      type:
        type: string
      updated_at:
        type: string
      user_name:
        type: string
    type: object
  models.StaffAuthResponse:
    properties:
      access_token:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      salon_id:
        type: string
      type:
        type: string
      username:
        type: string
    type: object
  models.StaffRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      image_url:
        type: string
      last_name:
        type: string
      passward:
        type: string
      phone_number:
        type: string
      salon_id:
        type: string
      type:
        enum:
        - superadmin
        - user
        type: string
      user_name:
        type: string
    required:
    - type
    type: object
  models.UpdatePasswordRequest:
    properties:
      password:
        type: string
    required:
    - password
    type: object
  models.UpdateStaffRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      image_url:
        type: string
      last_name:
        type: string
      passward:
        type: string
      phone_number:
        type: string
      user_name:
        type: string
    type: object
  models.VerifyRequest:
    properties:
      code:
        type: string
      email:
        type: string
    type: object
host: localhost:9090
info:
  contact: {}
  title: Swagger for barber shop api
  version: "1.0"
paths:
...
  /staff/create:
    post:
      consumes:
      - application/json
      description: This api for creating staff
      parameters:
      - description: Staff
        in: body
        name: staff
        required: true
        schema:
          $ref: '#/definitions/models.StaffRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Staff'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create staff
      tags:
      - staff
  /staff/delete/{id}:
    delete:
      consumes:
      - application/json
      description: This api for deleting staff by id
      parameters:
      - description: StaffID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseOK'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete staff by id
      tags:
      - staff
  /staff/forgot-password:
    post:
      consumes:
      - application/json
      description: This api for forgot password
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseOK'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: staff forgot password
      tags:
      - staff_auth
  /staff/get/{id}:
    get:
      consumes:
      - application/json
      description: This api for getting staff by id
      parameters:
      - description: StaffID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Staff'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get staff by id
      tags:
      - staff
  /staff/list:
    get:
      consumes:
      - application/json
      description: This api for getting list of staff
      parameters:
      - default: 10
        in: query
        name: limit
        required: true
        type: integer
      - default: 1
        in: query
        name: page
        required: true
        type: integer
      - in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetListStaffResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: get list staff
      tags:
      - staff
  /staff/login:
    post:
      consumes:
      - application/json
      description: This api for login staff
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.LogInRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.StaffAuthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: staff Login
      tags:
      - staff_auth
  /staff/register:
    post:
      consumes:
      - application/json
      description: This api for registeration staff
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.StaffRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseOK'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: staff register
      tags:
      - staff_auth
  /staff/update-password:
    post:
      consumes:
      - application/json
      description: This api for updating staff password
      parameters:
      - description: Password
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StaffAuthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: staff update password
      tags:
      - staff_auth
  /staff/update/{id}:
    put:
      consumes:
      - application/json
      description: This api for updating staff
      parameters:
      - description: StaffID
        in: path
        name: id
        required: true
        type: string
      - description: Staff
        in: body
        name: staff
        required: true
        schema:
          $ref: '#/definitions/models.UpdateStaffRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Staff'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a staff
      tags:
      - staff
  /staff/verify:
    post:
      consumes:
      - application/json
      description: This api for verification staff
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.VerifyRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StaffAuthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: staff verify
      tags:
      - staff_auth
  /staff/verify-forgot-password:
    post:
      consumes:
      - application/json
      description: Verify staff forgot password
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.VerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StaffAuthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: verify staff forgot password
      tags:
      - staff_auth
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"